# backend/docs/swagger.yaml
basePath: '{{.BasePath}}'
definitions:
  application.CourseReq:
    properties:
      code:
        type: integer
      description:
        type: string
      ementa:
        type: string
    required:
    - description
    - ementa
    type: object
  application.CourseStudentReq:
    properties:
      code_course:
        type: integer
      code_student:
        type: integer
    required:
    - code_course
    - code_student
    type: object
  application.StudentReq:
    properties:
      code:
        type: integer
      name:
        type: string
    required:
    - name
    type: object
  responses.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      message:
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Esse e um teste pratico para a vaga Flutter/Golang da empresa X.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TESTE PRATICO FLUTTER/GOLANG API
  version: 1.0.1
paths:
  /courses/add-students:
    post:
      description: Add students to course
      parameters:
      - description: Students and Course information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/application.CourseStudentReq'
          type: object
      - description: Timestamp of the request
        in: header
        name: Timestamp
        required: true
        type: string
      - description: Signature for the request
        in: header
        name: Sign
        required: true
        type: string
      - description: Unique UUID for the request
        in: header
        name: Nonce
        required: true
        type: string
      - description: API version
        in: header
        name: App-Version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Student add successfully
          schema:
            type: string
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/responses.Response'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.Response'
            type: object
      summary: Add students to course
      tags:
      - courses
  /courses/create:
    post:
      description: Create new course
      parameters:
      - description: Course information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/application.CourseReq'
          type: object
      - description: Timestamp of the request
        in: header
        name: Timestamp
        required: true
        type: string
      - description: Signature for the request
        in: header
        name: Sign
        required: true
        type: string
      - description: Unique UUID for the request
        in: header
        name: Nonce
        required: true
        type: string
      - description: API version
        in: header
        name: App-Version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Course created successfully
          schema:
            type: string
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/responses.Response'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.Response'
            type: object
      summary: Create course
      tags:
      - courses
  /courses/list:
    get:
      description: List all course
      parameters:
      - description: Limit the number of results (default is 10)
        in: query
        name: limit
        required: true
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        required: true
        type: integer
      - description: Course description
        in: query
        name: description
        type: string
      - description: Timestamp of the request
        in: header
        name: Timestamp
        required: true
        type: string
      - description: Signature for the request
        in: header
        name: Sign
        required: true
        type: string
      - description: Unique UUID for the request
        in: header
        name: Nonce
        required: true
        type: string
      - description: API version
        in: header
        name: App-Version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of courses
          schema:
            $ref: '#/definitions/responses.Response'
            type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/responses.Response'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.Response'
            type: object
      summary: List all course
      tags:
      - courses
  /courses/list-students-course:
    get:
      description: List all course students
      parameters:
      - description: Codigo do curso
        in: query
        name: search
        required: true
        type: integer
      - description: Timestamp of the request
        in: header
        name: Timestamp
        required: true
        type: string
      - description: Signature for the request
        in: header
        name: Sign
        required: true
        type: string
      - description: Unique UUID for the request
        in: header
        name: Nonce
        required: true
        type: string
      - description: API version
        in: header
        name: App-Version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of courses
          schema:
            $ref: '#/definitions/responses.Response'
            type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/responses.Response'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.Response'
            type: object
      summary: List all course students
      tags:
      - courses
  /courses/update:
    patch:
      description: Update course
      parameters:
      - description: Course information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/application.CourseReq'
          type: object
      - description: Timestamp of the request
        in: header
        name: Timestamp
        required: true
        type: string
      - description: Signature for the request
        in: header
        name: Sign
        required: true
        type: string
      - description: Unique UUID for the request
        in: header
        name: Nonce
        required: true
        type: string
      - description: API version
        in: header
        name: App-Version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Course updated successfully
          schema:
            type: string
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/responses.Response'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.Response'
            type: object
      summary: Update course
      tags:
      - courses
  /students/create:
    post:
      description: Create new student
      parameters:
      - description: Student information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/application.StudentReq'
          type: object
      - description: Timestamp of the request
        in: header
        name: Timestamp
        required: true
        type: string
      - description: Signature for the request
        in: header
        name: Sign
        required: true
        type: string
      - description: Unique UUID for the request
        in: header
        name: Nonce
        required: true
        type: string
      - description: API version
        in: header
        name: App-Version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Student created successfully
          schema:
            type: string
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/responses.Response'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.Response'
            type: object
      summary: Create student
      tags:
      - students
  /students/list:
    get:
      description: List all students
      parameters:
      - description: Limit the number of results (default is 10)
        in: query
        name: limit
        required: true
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        required: true
        type: integer
      - description: Student name
        in: query
        name: name
        type: string
      - description: Timestamp of the request
        in: header
        name: Timestamp
        required: true
        type: string
      - description: Signature for the request
        in: header
        name: Sign
        required: true
        type: string
      - description: Unique UUID for the request
        in: header
        name: Nonce
        required: true
        type: string
      - description: API version
        in: header
        name: App-Version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of students
          schema:
            $ref: '#/definitions/responses.Response'
            type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/responses.Response'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.Response'
            type: object
      summary: List all students
      tags:
      - students
  /students/update:
    patch:
      description: Update student
      parameters:
      - description: Student information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/application.StudentReq'
          type: object
      - description: Timestamp of the request
        in: header
        name: Timestamp
        required: true
        type: string
      - description: Signature for the request
        in: header
        name: Sign
        required: true
        type: string
      - description: Unique UUID for the request
        in: header
        name: Nonce
        required: true
        type: string
      - description: API version
        in: header
        name: App-Version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Student updated successfully
          schema:
            type: string
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/responses.Response'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.Response'
            type: object
      summary: Update student
      tags:
      - students
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"